test setup: stay away from the setgid mode bit
for vagrant testing on misc. platforms, we can't know the group /
we can't have the same group everywhere.

but the OS won't let us set setgid bit if the file does not have our group.
on netbsd, the created file somehow happens to have group "wheel",
but vagrant is not in group wheel.