Revert "use build_py to fix build on RTD", disables docs build
This reverts commit 8f0af3c309094f8a7206e7ab08710b28831cd3a8.

We will instead commit the generated `.rst` usage and API files
directly into git. The setup commands remain to generate them when the
usage or API changes, but as things are the hoops required to generate
those RST files are way too complicated to justify simply build docs.

See #384.

Conflicts:
	setup.py