fix PDF rendering structure
without those changes, all of the toctree document headings do not
show up. they are considered to be "below" the last heading of the
README file.

we also remove the "Notes" section from the readme as there is only
one note, regarding the fork.

we introduce a stub "introduction" element in the toctree, otherwise
it is impossible for the PDF rendered to render the README correctly.

this is to workaround a bug in the PDF renderer.