move fork differences to FAQ
it seems now that the fork is more of historical value than a current
thing. people interested in the differences between borg and attic can
look in the FAQ, but I do not see why this is present in the README.

a new section regarding compatibility is created to keep that warning
in place.