fix location parser for archives with @ char, add test, fixes #1930
we must exclude colon and slash chars from the username, otherwise the term for the user part
will match everything up to a @ char in the archive name.

a slash can't be in a username as the home directory would contain a illegal slash (slash is path sep),

a colon likely also should not be in a username because chown user:group ... syntax.